requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
python-multipart==0.0.9

streamlit==1.39.0

opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
pytesseract==0.3.13
pdf2image==1.17.0
scikit-image==0.24.0
